Changes to start procedures, kernel check, healthcheck, activation

+++ files/start.sh
@@ -1,11 +1,24 @@
-#!/usr/bin/bash
+#!/bin/bash
+
+if [[ -f "/etc/resolv.conf" ]]; then
 cp /etc/resolv.conf /etc/resolv.conf.bak
 umount /etc/resolv.conf
 cp /etc/resolv.conf.bak /etc/resolv.conf
 rm /etc/resolv.conf.bak
+fi
 sed -i 's/DAEMON_ARGS=.*/DAEMON_ARGS=""/' /etc/init.d/expressvpn
-service expressvpn restart
-expect /expressvpn/activate.sh
+output=$(service expressvpn restart 2>&1)
+if [[ $output == *"failed!"* ]]; then
+    echo "Service expressvpn restart failed!"
+    bash /expressvpn/start.sh
+    exit 1
+fi
+output=$(expect /expressvpn/activate.sh 2>&1)
+if [[ $output == *"Please activate your account."* ]]; then
+    echo "Activation failed!"
+    bash /expressvpn/start.sh
+    exit 1
+fi
 expressvpn preferences set preferred_protocol $PROTOCOL
 expressvpn preferences set lightway_cipher $CIPHER
 expressvpn preferences set send_diagnostics false
